Add view a point button to landing page
